requirements.txt
africastalking==1.2.8
altgraph==0.17.4
asgiref==3.8.1
attrs==24.2.0
Authlib==1.3.2
bleach==6.1.0
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
cryptography==43.0.1
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==4.2.16
django-cors-headers==4.4.0
django-heroku==0.3.1
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
ecdsa==0.19.0
exceptiongroup==1.2.2
Flask==3.0.3
future==1.0.0
greenlet==3.1.0
gunicorn==23.0.0
h11==0.14.0
idna==3.8
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jwcrypto==1.5.6
macholib==1.16.3
MarkupSafe==2.1.5
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
psycopg2==2.9.9
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.9.0
python-jose==3.3.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schema==0.7.7
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sortedcontainers==2.4.0
SQLAlchemy==2.0.34
sqlparse==0.5.1
tomli==2.0.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
whitenoise==6.7.0
wsproto==1.2.0
zipp==3.20.2
